+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- "More" submenu with Undo, Beverage settings, and Reset counter options
+
 ---
 
 ## [0.12] - 2026-01-17